feat: add function to check for necessary dependencies in install.sh
A new function 'check_top_level_dependencies' has been added to the install.sh script.
This function checks for the existence of all top-level dependencies.
If any required dependencies are missing, the function alerts the user and provides instructions
for installation on different operating systems.
The check for top-level dependencies is now being conducted at the start of the installation
process.