feat: use app token  to work around branch protection during the semantic release workflow
Updates the Semantic Release workflow to include a workaround for pushing tags to a protected
branch.
Adds a step to create a GitHub App token using the `actions/create-github-app-token` action.
Updates the checkout action to use the generated token and fetches the full git history.
Adds an additional usage of the generated app token in the Semantic Release run step.
Comments out the previously-used credentials in favor of the app token.